Move TLS related routines to go/db/tls.go (related to PR#139)
