+++ go/db/db.go
@@ -17,14 +17,11 @@
 package db
 
 import (
-	"crypto/tls"
 	"database/sql"
 	"fmt"
 	"strings"
 
 	"github.com/github/orchestrator/go/config"
-	"github.com/github/orchestrator/go/ssl"
-	"github.com/go-sql-driver/mysql"
 	"github.com/openark/golib/log"
 	"github.com/openark/golib/sqlutils"
 )
@@ -1227,12 +1224,6 @@ var generateSQLPatches = []string{
 	`,
 }
 
-// Track if a TLS has already been configured for topology
-var topologyTLSConfigured bool = false
-
-// Track if a TLS has already been configured for Orchestrator
-var orchestratorTLSConfigured bool = false
-
 // OpenDiscovery returns a DB instance to access a topology instance.
 // It has lower read timeout than OpenTopology and is intended to
 // be used with low-latency discovery queries.
@@ -1260,30 +1251,6 @@ func openTopology(host string, port int, readTimeout int) (*sql.DB, error) {
 	db.SetMaxOpenConns(config.Config.MySQLTopologyMaxPoolConnections)
 	db.SetMaxIdleConns(config.Config.MySQLTopologyMaxPoolConnections)
 	return db, err
-}
-
-// Create a TLS configuration from the config supplied CA, Certificate, and Private key.
-// Register the TLS config with the mysql drivers as the "topology" config
-// Modify the supplied URI to call the TLS config
-// TODO: Way to have password mixed with TLS for various nodes in the topology.  Currently everything is TLS or everything is password
-func SetupMySQLTopologyTLS(uri string) (string, error) {
-	if !topologyTLSConfigured {
-		tlsConfig, err := ssl.NewTLSConfig(config.Config.MySQLTopologySSLCAFile, !config.Config.MySQLTopologySSLSkipVerify)
-		// Drop to TLS 1.0 for talking to MySQL
-		tlsConfig.MinVersion = tls.VersionTLS10
-		if err != nil {
-			return "", log.Fatalf("Can't create TLS configuration for Topology connection %s: %s", uri, err)
-		}
-		tlsConfig.InsecureSkipVerify = config.Config.MySQLTopologySSLSkipVerify
-		if err = ssl.AppendKeyPair(tlsConfig, config.Config.MySQLTopologySSLCertFile, config.Config.MySQLTopologySSLPrivateKeyFile); err != nil {
-			return "", log.Fatalf("Can't setup TLS key pairs for %s: %s", uri, err)
-		}
-		if err = mysql.RegisterTLSConfig("topology", tlsConfig); err != nil {
-			return "", log.Fatalf("Can't register mysql TLS config for topology: %s", err)
-		}
-		topologyTLSConfigured = true
-	}
-	return fmt.Sprintf("%s&tls=topology", uri), nil
 }
 
 // OpenTopology returns the DB instance for the orchestrator backed database
@@ -1370,29 +1337,6 @@ func registerOrchestratorDeployment(db *sql.DB) error {
 	}
 	log.Debugf("Migrated database schema to version [%+v]", config.RuntimeCLIFlags.ConfiguredVersion)
 	return nil
-}
-
-// Create a TLS configuration from the config supplied CA, Certificate, and Private key.
-// Register the TLS config with the mysql drivers as the "orchestrator" config
-// Modify the supplied URI to call the TLS config
-func SetupMySQLOrchestratorTLS(uri string) (string, error) {
-	if !orchestratorTLSConfigured {
-		tlsConfig, err := ssl.NewTLSConfig(config.Config.MySQLOrchestratorSSLCAFile, true)
-		// Drop to TLS 1.0 for talking to MySQL
-		tlsConfig.MinVersion = tls.VersionTLS10
-		if err != nil {
-			return "", log.Fatalf("Can't create TLS configuration for Orchestrator connection %s: %s", uri, err)
-		}
-		tlsConfig.InsecureSkipVerify = config.Config.MySQLOrchestratorSSLSkipVerify
-		if err = ssl.AppendKeyPair(tlsConfig, config.Config.MySQLOrchestratorSSLCertFile, config.Config.MySQLOrchestratorSSLPrivateKeyFile); err != nil {
-			return "", log.Fatalf("Can't setup TLS key pairs for %s: %s", uri, err)
-		}
-		if err = mysql.RegisterTLSConfig("orchestrator", tlsConfig); err != nil {
-			return "", log.Fatalf("Can't register mysql TLS config for orchestrator: %s", err)
-		}
-		orchestratorTLSConfigured = true
-	}
-	return fmt.Sprintf("%s&tls=orchestrator", uri), nil
 }
 
 // deployStatements will issue given sql queries that are not already known to be deployed.
